requirements.txt
tensorflow==2.13.0
numpy==1.24.4
pillow==9.5.0
streamlit==1.26.0